Move Charge out of charge_types
The charge_types/ folder should only contain subclasses on Charge. This will also make it easier to dynamically list out all subclasses of Charge by iterating through the folder without having to check for the exception (charge.py).
